test: mailbox, igp, ism, announce (#25)

+++ x/core/keeper/logic_ism.go
@@ -33,7 +33,6 @@ func (k Keeper) Verify(ctx context.Context, ismId util.HexAddress, rawMetadata [
 
 	switch v := ism.Ism.(type) {
 	case *types.Ism_MultiSig:
-
 		metadata, err := types.NewMetadata(rawMetadata)
 		if err != nil {
 			return false, err
@@ -57,7 +56,7 @@ func (k Keeper) Verify(ctx context.Context, ismId util.HexAddress, rawMetadata [
 		}
 
 		signatures, validSignatures := metadata.SignatureCount(), uint32(0)
-		threshold := uint32(multiSigIsm.Threshold)
+		threshold := multiSigIsm.Threshold
 
 		// Early return if we can't possibly meet the threshold
 		if signatures < multiSigIsm.Threshold {
